# TCCPOS.Backend.InventoryService.WebApi/InventoryServiceSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Inventory WebApi",
    "description": "TCCPOS Backend Inventory WebApi",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Group": {
      "post": {
        "tags": [
          "Group"
        ],
        "summary": "Create Group",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopGroupRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopGroupRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateShopGroupCommand"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateShopGroupCommand"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateShopGroupCommand"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Group"
        ],
        "summary": "Get all group",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllMerchantGroupResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllMerchantGroupResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllMerchantGroupResult"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Group/{id}": {
      "get": {
        "tags": [
          "Group"
        ],
        "summary": "Get group by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetMerchantGroupByIdResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMerchantGroupByIdResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMerchantGroupByIdResult"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Group"
        ],
        "summary": "Delete group by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteShopGroupResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteShopGroupResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteShopGroupResult"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Group/Detail": {
      "put": {
        "tags": [
          "Group"
        ],
        "summary": "Update group detail",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGroupResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGroupResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGroupResult"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Group/Name": {
      "put": {
        "tags": [
          "Group"
        ],
        "summary": "Update group name",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupNameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupNameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupNameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGroupNameResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGroupNameResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGroupNameResult"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Group/Target/{id}": {
      "get": {
        "tags": [
          "Group"
        ],
        "summary": "Get target group by id",
        "description": "",
        "parameters": [
          {
            "name": "shopgroupid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantGroupResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantGroupResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantGroupResult"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Merchant/GetAllShopWithAddress": {
      "get": {
        "tags": [
          "Merchant"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllMerchantAddressResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllMerchantAddressResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllMerchantAddressResult"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Merchant/getAllShop": {
      "get": {
        "tags": [
          "Merchant"
        ],
        "summary": "",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllShopResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllShopResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllShopResult"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateShopGroupCommand": {
        "type": "object",
        "properties": {
          "shopGroupName": {
            "type": "string",
            "nullable": true
          },
          "shopId": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateShopGroupRequest": {
        "required": [
          "shopGroupName",
          "shopId"
        ],
        "type": "object",
        "properties": {
          "shopGroupName": {
            "minLength": 1,
            "type": "string"
          },
          "shopId": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "DeleteShopGroupResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FailedResult": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "errorDetail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetAllMerchantAddressResult": {
        "type": "object",
        "properties": {
          "shopAddress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShopWithAddressResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetAllMerchantGroupResult": {
        "type": "object",
        "properties": {
          "shopGroupId": {
            "type": "string",
            "nullable": true
          },
          "shopGroupName": {
            "type": "string",
            "nullable": true
          },
          "shopAmount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "GetAllShopResult": {
        "type": "object",
        "properties": {
          "shopId": {
            "type": "string",
            "nullable": true
          },
          "shopName": {
            "type": "string",
            "nullable": true
          },
          "shopGroupId": {
            "type": "string",
            "nullable": true
          },
          "shopGroupName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetMerchantGroupByIdResult": {
        "type": "object",
        "properties": {
          "shop_group_id": {
            "type": "string",
            "nullable": true
          },
          "group_name": {
            "type": "string",
            "nullable": true
          },
          "shopList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShopResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MerchantGroupResult": {
        "type": "object",
        "properties": {
          "targetID": {
            "type": "string",
            "nullable": true
          },
          "skuID": {
            "type": "string",
            "nullable": true
          },
          "skuName": {
            "type": "string",
            "nullable": true
          },
          "target": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rewardID": {
            "type": "string",
            "nullable": true
          },
          "reward": {
            "type": "string",
            "nullable": true
          },
          "resetDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShopResult": {
        "type": "object",
        "properties": {
          "shopId": {
            "type": "string",
            "nullable": true
          },
          "shopName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ShopWithAddressResult": {
        "type": "object",
        "properties": {
          "shop_id": {
            "type": "string",
            "nullable": true
          },
          "shop_name": {
            "type": "string",
            "nullable": true
          },
          "shop_address_id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateGroupNameRequest": {
        "required": [
          "shopGroupId",
          "shopGroupName"
        ],
        "type": "object",
        "properties": {
          "shopGroupId": {
            "minLength": 1,
            "type": "string"
          },
          "shopGroupName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateGroupNameResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateGroupRequest": {
        "required": [
          "shopGroupId",
          "shopGroupName",
          "shopList"
        ],
        "type": "object",
        "properties": {
          "shopGroupId": {
            "minLength": 1,
            "type": "string"
          },
          "shopGroupName": {
            "minLength": 1,
            "type": "string"
          },
          "shopList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "UpdateGroupResult": {
        "type": "object",
        "properties": {
          "shopgroup": {
            "$ref": "#/components/schemas/shopgroup"
          },
          "shops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/shop"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "shop": {
        "type": "object",
        "properties": {
          "shop_id": {
            "type": "string",
            "nullable": true
          },
          "shop_name": {
            "type": "string",
            "nullable": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_by": {
            "type": "string",
            "nullable": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_by": {
            "type": "string",
            "nullable": true
          },
          "price_tier_id": {
            "type": "string",
            "nullable": true
          },
          "shop_group_id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "shopgroup": {
        "type": "object",
        "properties": {
          "shop_group_id": {
            "type": "string",
            "nullable": true
          },
          "group_name": {
            "type": "string",
            "nullable": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_by": {
            "type": "string",
            "nullable": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_by": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}